+++ src/vue/modules/requests/create-asset-requests/components/requests-table.vue
@@ -1,7 +1,6 @@
 <template>
   <div class="requests-table">
     <div
-      v-if="requests.length"
       class="app__table app__table--with-shadow"
     >
       <table>
@@ -24,7 +23,9 @@
           </tr>
         </thead>
 
-        <tbody>
+        <tbody
+          v-if="requests.length"
+        >
           <tr
             v-for="request in requests"
             :key="request.id"
@@ -55,21 +56,19 @@
             </td>
           </tr>
         </tbody>
+        <empty-list-placeholder
+          v-else
+          :colspan="5"
+          :message="'create-asset-requests.no-request-history-desc' | globalize"
+        />
       </table>
     </div>
-
-    <no-data-message
-      v-else
-      icon-name="trending-up"
-      :title="'create-asset-requests.no-request-history-title' | globalize"
-      :message="'create-asset-requests.no-request-history-desc' | globalize"
-    />
   </div>
 </template>
 
 <script>
-import NoDataMessage from '@/vue/common/NoDataMessage'
 import RequestStateViewer from '../../shared/components/request-state-viewer'
+import EmptyListPlaceholder from '@/vue/common/EmptyListPlaceholder'
 
 const EVENTS = {
   select: 'select',
@@ -78,8 +77,8 @@ const EVENTS = {
 export default {
   name: 'requests-table',
   components: {
-    NoDataMessage,
     RequestStateViewer,
+    EmptyListPlaceholder,
   },
 
   props: {